+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New functions
 
 * `impute_in_classes()` allows to apply any imputation function inside imputation classes 
+* `impute_hot_deck_in_classes()` hot deck imputation inside of imputation classes (adjustment cells)
 
 ## Miscellaneous